Crop each video frame as we save it
By doing this, we can adjust the crop area to fix the one column
offset to the left it previously had in MODE7.  This is due to
features of the real hardware which jsbeeb emulates (see
https://github.com/mattgodbolt/jsbeeb/issues/301).

This also nearly halves the temporary disk space needed to save
the video frames and speeds up the bot by a few percent.

This changes all the checksums in the testcases because they're
now calculated from the cropped final frame instead of the full final
frame.  I've verified by comparing output before and after (for the
non-MODE7 testcases, the output media is bit-for-bit identical).