Trim explicit "RUN" from program with line numbers
With beebjit we're using the tokeniser for everything, and this
causes such a "RUN" to be given a line number and added to the
program, which is unlikely to be what was wanted.  Some older
examples re-sent to the bot behave differently due to this.

Fixes testcase "YOUONLYRUNONCE" (the checksum update is needed
because beebjit produces bgra, and has some minor MODE7 rendering
differences compared to jsbeeb).