Drop NOVSYNC testcase
This test was added as a regression test for the case where no
frames of output are produced, but current versions of both jsbeeb
and beebjit now produce output so it isn't actually useful for what
it was added for, and it isn't really a useful test apart from its
original purpose.

I can't see how to get current beebjit to generate no frames, so
just stop trying to test this obscure corner case.