Don't load DFS by default
DFS was an add-on to the original hardware and isn't really useful with
the bot.  Loading it by default raises PAGE by 2816 bytes.  Even in the
best case of MODE7 and rocket mode, that's a 10% reduction in the
available RAM to BBC BASIC.  In MODE2 and with the GXR loaded it's a
whopping 35% reduction.

Support enabling DFS via an emoji, which is probably only really useful
as a way to allow easily running programs written for older versions of
the bot which assume the value of PAGE or the screen contents.