Drop support for no-entry emoji
Programs for the bot to run are now neatly separated from discussion
with the bot operator, since the former now need to have a #bbcmicrobot
hashtag while the latter shouldn't have one (and experience so far
is that people understand this).

This means the no-entry emoji is no longer useful.  There's no reason
to use the hashtag on a conversation, but if someone fails to understand
that and uses it anyway, they're not going to know to include a magic
emoji in the message to stop the bot running it.

This emoji was also never actually documented in the README or user
guide, and didn't work properly in the twitter incarnation of the bot
(based on seeing a few people try to use it).